Fri, May 27, 2022, 3:23 AM +00:00

+++ Week 0/pennies.c
@@ -5,7 +5,7 @@
 int main(void)
 {
     float amount = get_float("Dollar amount: ");
-    float pennies = round(amount * 100);
+    int pennies = round(amount * 100);
 
-    printf("Pennies: %f\n", pennies);
+    printf("Pennies: %i\n", pennies);
 }
